Rename main to index, fix issues/add enhancements
